Make all implementations of Mul and Div traits explicit.
The reason for this is that due to (what I assume is) a bug in the
compiler when using (unstable) generic_const_exprs, trait
implementations that were too generic would result in an overflow while
the compiler was presumably looking for candidates of some Div
operation. This was hard to reproduce and only happened in a random
place in raxiom, where `ndarray` is in scope, which itself implements a
few Div and Mul traits with floats, complicating the situation.