Add `#![feature(generic_const_exprs)]` to doctests
Some of them were missing this feature annotation but used to compile
and pass for some inexplicable reason. Now this missing feature gate
causes an ICE (https://github.com/rust-lang/rust/issues/133199). This
adds the feature gates in all the doctests, but does not yet address the
fact that this ICE also occurs in calling crates that don't have
`#![feature(generic_const_exprs)]` set. Maybe a workaround could be to
check for this feature explicitly and throw a helpful error message in
diman.